before rename UsageLog

+++ VpnHood.AccessServer/recreate-db.ps1
@@ -1,4 +1,9 @@
-Remove-Item Migrations\* -Recurse
-dotnet ef database drop -f
-dotnet ef migrations add Init
-dotnet ef database update
+$migrationPath="$PSScriptRoot/Migrations";
+if (Test-Path $migrationPath )
+{
+	Remove-Item Migrations\* -Recurse;
+}
+
+dotnet ef database drop -f;
+dotnet ef migrations add Init;
+dotnet ef database update;
